+++ src/modules/transactions/components/TransactionCard/transfer-details/address-actions/options/copy-address.tsx
